libnsl: cast between incompatible function types
../rpc/xdr_sizeof.c: In function 'xdr_sizeof':
../rpc/xdr_sizeof.c:136:19: error: cast between incompatible function types from
 'int (*)(void)' to 'bool_t (*)(XDR *, long int *)' {aka 'int (*)(struct XDR *,
long int *)'} [-Werror=cast-function-type]
  ops.x_getlong =  (dummyfunc1) harmless;
                   ^
../rpc/xdr_sizeof.c:148:19: error: cast between incompatible function types from
 'int (*)(void)' to 'bool_t (*)(XDR *, char *, int)' {aka 'int (*)(struct XDR *,
 char *, int)'} [-Werror=cast-function-type]
  ops.x_getbytes = (dummyfunc2) harmless;
                   ^
cc1: all warnings being treated as errors

+++ usr/src/lib/libnsl/rpc/xdr_sizeof.c
@@ -133,19 +133,19 @@ xdr_sizeof(xdrproc_t func, void *data)
 #endif
 
 	ops.x_putlong = x_putlong;
-	ops.x_getlong =  (dummyfunc1) harmless;
+	ops.x_getlong =  (dummyfunc1)(uintptr_t) harmless;
 	ops.x_putbytes = x_putbytes;
 	ops.x_inline = x_inline;
 	ops.x_getpostn = x_getpostn;
 	ops.x_setpostn = x_setpostn;
 	ops.x_destroy = x_destroy;
 #if defined(_LP64)
-	ops.x_getint32 = (dummyfunc3) harmless;
+	ops.x_getint32 = (dummyfunc3)(uintptr_t) harmless;
 	ops.x_putint32 = x_putint32_t;
 #endif
 
 	/* the other harmless ones */
-	ops.x_getbytes = (dummyfunc2) harmless;
+	ops.x_getbytes = (dummyfunc2)(uintptr_t) harmless;
 
 	x.x_op = XDR_ENCODE;
 	x.x_ops = &ops;
